+++ ssvsigner/web3signer/web3signer.go
@@ -7,11 +7,19 @@ import (
 	"strings"
 	"time"
 
+	v1 "github.com/attestantio/go-eth2-client/api/v1"
+	"github.com/attestantio/go-eth2-client/spec/altair"
 	"github.com/attestantio/go-eth2-client/spec/phase0"
 	"github.com/carlmjohnson/requests"
 )
 
 const DefaultRequestTimeout = 10 * time.Second
+
+const (
+	pathPublicKeys = "/api/v1/eth2/publicKeys"
+	pathKeystores  = "/eth/v1/keystores"
+	pathSign       = "/api/v1/eth2/sign/"
+)
 
 type Web3Signer struct {
 	baseURL    string
@@ -43,6 +51,8 @@ func WithRequestTimeout(timeout time.Duration) Option {
 	}
 }
 
+type ListKeysResponse []phase0.BLSPubKey
+
 // ListKeys lists keys in Web3Signer using https://consensys.github.io/web3signer/web3signer-eth2.html#tag/Public-Key/operation/ETH2_LIST
 func (w3s *Web3Signer) ListKeys(ctx context.Context) (ListKeysResponse, error) {
 	var resp ListKeysResponse
@@ -53,6 +63,17 @@ func (w3s *Web3Signer) ListKeys(ctx context.Context) (ListKeysResponse, error) {
 		ToJSON(&resp).
 		Fetch(ctx)
 	return resp, w3s.handleWeb3SignerErr(err)
+}
+
+type ImportKeystoreRequest struct {
+	Keystores          []string `json:"keystores"`
+	Passwords          []string `json:"passwords"`
+	SlashingProtection string   `json:"slashing_protection,omitempty"`
+}
+
+type ImportKeystoreResponse struct {
+	Data    []KeyManagerResponseData `json:"data,omitempty"`
+	Message string                   `json:"message,omitempty"`
 }
 
 // ImportKeystore adds a key to Web3Signer using https://consensys.github.io/web3signer/web3signer-eth2.html#tag/Keymanager/operation/KEYMANAGER_IMPORT
@@ -69,6 +90,16 @@ func (w3s *Web3Signer) ImportKeystore(ctx context.Context, req ImportKeystoreReq
 	return resp, w3s.handleWeb3SignerErr(err)
 }
 
+type DeleteKeystoreRequest struct {
+	Pubkeys []phase0.BLSPubKey `json:"pubkeys"`
+}
+
+type DeleteKeystoreResponse struct {
+	Data               []KeyManagerResponseData `json:"data,omitempty"`
+	SlashingProtection string                   `json:"slashing_protection,omitempty"`
+	Message            string                   `json:"message,omitempty"`
+}
+
 // DeleteKeystore removes a key from Web3Signer using https://consensys.github.io/web3signer/web3signer-eth2.html#operation/KEYMANAGER_DELETE
 func (w3s *Web3Signer) DeleteKeystore(ctx context.Context, req DeleteKeystoreRequest) (DeleteKeystoreResponse, error) {
 	var resp DeleteKeystoreResponse
@@ -81,6 +112,26 @@ func (w3s *Web3Signer) DeleteKeystore(ctx context.Context, req DeleteKeystoreReq
 		ToJSON(&resp).
 		Fetch(ctx)
 	return resp, w3s.handleWeb3SignerErr(err)
+}
+
+type SignRequest struct {
+	ForkInfo                    ForkInfo                          `json:"fork_info"`
+	SigningRoot                 phase0.Root                       `json:"signing_root,omitempty"`
+	Type                        SignedObjectType                  `json:"type"`
+	Attestation                 *phase0.AttestationData           `json:"attestation,omitempty"`
+	BeaconBlock                 *BeaconBlockData                  `json:"beacon_block,omitempty"`
+	VoluntaryExit               *phase0.VoluntaryExit             `json:"voluntary_exit,omitempty"`
+	AggregateAndProof           *AggregateAndProof                `json:"aggregate_and_proof,omitempty"`
+	AggregationSlot             *AggregationSlot                  `json:"aggregation_slot,omitempty"`
+	RandaoReveal                *RandaoReveal                     `json:"randao_reveal,omitempty"`
+	SyncCommitteeMessage        *SyncCommitteeMessage             `json:"sync_committee_message,omitempty"`
+	SyncAggregatorSelectionData *SyncCommitteeAggregatorSelection `json:"sync_aggregator_selection_data,omitempty"`
+	ContributionAndProof        *altair.ContributionAndProof      `json:"contribution_and_proof,omitempty"`
+	ValidatorRegistration       *v1.ValidatorRegistration         `json:"validator_registration,omitempty"`
+}
+
+type SignResponse struct {
+	Signature phase0.BLSSignature `json:"signature"`
 }
 
 // Sign signs using https://consensys.github.io/web3signer/web3signer-eth2.html#tag/Signing/operation/ETH2_SIGN
